Responsive video embeds and video section
You can now use the `video:` frontmatter to add the embed code to an
episode. It will put it in it’s own section called “Video” with a width
of 100% at a 16:9 ratio. You’ll need to copy the whole embed code (any
size, default is fine) and paste it into the `video:` section for the
episode.